- fixed a few warnings

+++ src/RenderSystem.cpp
@@ -203,8 +203,8 @@ namespace april
 	void RenderSystem::setOrthoProjection(grect rect)
 	{
 		float t = getPixelOffset();
-		float wnd_w = mWindow->getWidth();
-		float wnd_h = mWindow->getHeight();
+		float wnd_w = (float)mWindow->getWidth();
+		float wnd_h = (float)mWindow->getHeight();
 		mProjectionMatrix.ortho(rect.w, rect.h, -rect.x + t * rect.w / wnd_w, -rect.y + t * rect.h / wnd_h);
 		_setProjectionMatrix(mProjectionMatrix);
 	}
